refactor: remove wrongfully added file while merging
